requirements.txt
argparse
faiss-cpu
itertools
langchain
langchain-openai
PyPDF2
pypdf
python-dotenv
langchain-text-splitters
langchain-openai faiss-cpu